requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
httpx>=0.24.0
numpy>=1.24.3
pandas>=2.0.0
faiss-cpu>=1.7.4
openai>=0.27.0
pydantic>=2.0.0